requirements.txt
absl-py==0.7.1
astor==0.8.0
blessings==1.7
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
gpustat==0.6.0
grpcio==1.22.0
h5py==2.9.0
imageio==2.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.16.3
nvidia-ml-py3==7.352.0
pandas==0.25.0
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.9.0
psutil==5.6.3
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
scikit-image==0.15.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.2
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
